ESF AUTH_EXEC: offload the hashing decision off the serial delivery thread (#298)
handleAuthExec ran the whole decision — SecCode walk + SHA-256 + kernel
respond — synchronously on the ES client's single serial AUTH delivery thread.
One slow hash (cold cache + BINARY rules, e.g. the two-client window of an
in-place extension upgrade) serialized every queued exec, and a queued exec
that burned its kernel deadline while waiting was fail-closed DENIED. The
result was a system-wide exec stall with spurious denials (the #298 incident
on edr-qa rc.2).

Fix: keep the two instant, no-I/O fast paths (platform-binary carve-out,
self-allow failsafe) inline, and offload everything past them to a bounded
OperationQueue (authWorkerConcurrency = 4). es_retain_message keeps the message
valid past the callback; the worker walks the tuple, hashes, decides, responds,
and es_release_message's it. The ES thread returns immediately, so a slow hash
can no longer serialize other execs. stop() drains in-flight workers before
es_delete_client.

Decision logic, the msg.deadline budget, and the fail-closed/open posture are
unchanged — only the thread they run on moves. All shared state the worker
touches (FileHashCache, SigningInfoFallback, the snapshot, the serializer, the
XPC send) is already OSAllocatedUnfairLock-guarded / pure / serialized, so no
new locking is needed. The grace-window / failClosed-default band-aids from the
issue are deliberately omitted: they weaken enforcement, and the offload makes
them unnecessary. Cache-warming across upgrades is a separate change.

buildAuthTuple + the emit* helpers move to ESFSubscriber+AuthExec.swift to keep
ESFSubscriber.swift under SwiftLint's file_length / type_body_length caps (same
split rationale as +FileEvents / +BTM).
